backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.0
django-cors-headers==4.8.0
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
gunicorn==21.2.0